refactor: move test_workspace fixture to conftest.py to reuse
